tf.batch_matrix_diag_part=>tf.matrix_diag_part()

+++ Chapter 03/logistic_regression.py
@@ -32,7 +32,7 @@ y_vals = np.array([x[1] for x in birth_data])
 x_vals = np.array([x[2:9] for x in birth_data])
 
 # Split data into train/test = 80%/20%
-train_indices = np.random.choice(len(x_vals), round(len(x_vals)*0.8), replace=False)
+train_indices = np.random.choice(len(x_vals), int(round(len(x_vals)*0.8)), replace=False)
 test_indices = np.array(list(set(range(len(x_vals))) - set(train_indices)))
 x_vals_train = x_vals[train_indices]
 x_vals_test = x_vals[test_indices]
